Minor documentation changes, data struct renamed
I forgot to change the names of some of the struct in the myoapi header,
this is now corrected. These structures are internal so there are no
changes to the API. All of the internal structures where also added to
the internal group to indicate they are only used internally.